Hex print pointers without first casting to `usize`

+++ hook/src/hooks/user/pawn.rs
@@ -57,7 +57,7 @@ impl Pawns {
             common::log!("Tracking no pawns.");
         } else {
             for (i, pawn) in self.pawns.iter().enumerate() {
-                common::log!("[{}] {} {}", i, common::Hex(pawn.pointer as usize), (*pawn.pointer.cast::<common::UObject>()).name());
+                common::log!("[{}] {} {}", i, common::Hex(pawn.pointer), (*pawn.pointer.cast::<common::UObject>()).name());
             }
         }
     }
